policies: rewrite cache_key to prevent wrong cache
