+++ passbook/policies/process.py
@@ -14,11 +14,13 @@ from passbook.policies.types import PolicyRequest, PolicyResult
 LOGGER = get_logger()
 
 
-def cache_key(policy: Policy, user: Optional[User] = None) -> str:
+def cache_key(policy: Policy, request: PolicyRequest) -> str:
     """Generate Cache key for policy"""
     prefix = f"policy_{policy.pk}"
-    if user:
-        prefix += f"#{user.pk}"
+    if request.http_request:
+        prefix += f"_{request.http_request.session.session_key}"
+    if request.user:
+        prefix += f"#{request.user.pk}"
     return prefix
 
 
@@ -59,7 +61,7 @@ class PolicyProcess(Process):
             passing=policy_result.passing,
             user=self.request.user,
         )
-        key = cache_key(self.policy, self.request.user)
+        key = cache_key(self.policy, self.request)
         cache.set(key, policy_result)
         LOGGER.debug("P_ENG(proc): Cached policy evaluation", key=key)
         self.connection.send(policy_result)
